stock_news_pipeline/airflow/requirements.txt
yfinance
multitasking<0.0.11
pandas
requests
beautifulsoup4
psycopg2-binary
python-dotenv
feedparser